requirements.txt
Django>=2.2.7
django-nested-inline>=0.3.7